+++ Sources/TuistGraph/Models/Plugins.swift
@@ -9,22 +9,29 @@ public struct Plugins: Equatable {
     /// List of paths to template definitions.
     public let templateDirectories: [AbsolutePath]
 
+    /// List of paths pointing to resource templates
+    public let resourceSynthesizers: [PluginResourceSynthesizer]
+
     /// Creates a `Plugins`.
     ///
     /// - Parameters:
     ///     - projectDescriptionHelpers: List of the loaded helper plugins.
     ///     - templatePaths: List of paths to the `Templates/` directory for the loaded plugins.
+    ///     - resourceSynthesizers: List of the loaded resource synthesizer plugins
     public init(
         projectDescriptionHelpers: [ProjectDescriptionHelpersPlugin],
-        templatePaths: [AbsolutePath]
+        templatePaths: [AbsolutePath],
+        resourceSynthesizers: [PluginResourceSynthesizer]
     ) {
         self.projectDescriptionHelpers = projectDescriptionHelpers
         templateDirectories = templatePaths
+        self.resourceSynthesizers = resourceSynthesizers
     }
 
     /// An empty `Plugins`.
     public static let none: Plugins = .init(
         projectDescriptionHelpers: [],
-        templatePaths: []
+        templatePaths: [],
+        resourceSynthesizers: []
     )
 }