+++ Sources/TuistGraph/Models/ResourceSynthesizer.swift
@@ -1,0 +1,34 @@
+import Foundation
+import TSCBasic
+
+public struct ResourceSynthesizer: Equatable, Hashable {
+    public let parser: Parser
+    public let extensions: Set<String>
+    public let template: Template
+
+    public enum Template: Equatable, Hashable {
+        case file(AbsolutePath)
+        case defaultTemplate(String)
+    }
+
+    public enum Parser: Equatable, Hashable {
+        case strings
+        case assets
+        case plists
+        case fonts
+        case coreData
+        case interfaceBuilder
+        case json
+        case yaml
+    }
+
+    public init(
+        parser: Parser,
+        extensions: Set<String>,
+        template: Template
+    ) {
+        self.parser = parser
+        self.extensions = extensions
+        self.template = template
+    }
+}