+++ Sources/TuistLoader/Loaders/ManifestModelConverter.swift
@@ -8,11 +8,16 @@ import TuistSupport
 /// A component responsible for converting Manifests (`ProjectDescription`) to Models (`TuistCore`)
 public protocol ManifestModelConverting {
     func convert(manifest: ProjectDescription.Workspace, path: AbsolutePath) throws -> TuistGraph.Workspace
-    func convert(manifest: ProjectDescription.Project, path: AbsolutePath) throws -> TuistGraph.Project
+    func convert(
+        manifest: ProjectDescription.Project,
+        path: AbsolutePath,
+        plugins: Plugins
+    ) throws -> TuistGraph.Project
 }
 
 public final class ManifestModelConverter: ManifestModelConverting {
     private let manifestLoader: ManifestLoading
+    private let resourceSynthesizerPathLocator: ResourceSynthesizerPathLocating
 
     public convenience init() {
         self.init(
@@ -20,16 +25,35 @@ public final class ManifestModelConverter: ManifestModelConverting {
         )
     }
 
-    public init(manifestLoader: ManifestLoading) {
+    public convenience init(
+        manifestLoader: ManifestLoading
+    ) {
+        self.init(
+            manifestLoader: manifestLoader,
+            resourceSynthesizerPathLocator: ResourceSynthesizerPathLocator()
+        )
+    }
+
+    init(
+        manifestLoader: ManifestLoading,
+        resourceSynthesizerPathLocator: ResourceSynthesizerPathLocating = ResourceSynthesizerPathLocator()
+    ) {
         self.manifestLoader = manifestLoader
+        self.resourceSynthesizerPathLocator = resourceSynthesizerPathLocator
     }
 
     public func convert(
         manifest: ProjectDescription.Project,
-        path: AbsolutePath
+        path: AbsolutePath,
+        plugins: Plugins
     ) throws -> TuistGraph.Project {
         let generatorPaths = GeneratorPaths(manifestDirectory: path)
-        return try TuistGraph.Project.from(manifest: manifest, generatorPaths: generatorPaths)
+        return try TuistGraph.Project.from(
+            manifest: manifest,
+            generatorPaths: generatorPaths,
+            plugins: plugins,
+            resourceSynthesizerPathLocator: resourceSynthesizerPathLocator
+        )
     }
 
     public func convert(