+++ Tests/TuistIntegrationTests/Generator/MultipleConfigurationsIntegrationTests.swift
@@ -377,7 +377,8 @@ final class MultipleConfigurationsIntegrationTests: TuistUnitTestCase {
             packages: packages,
             schemes: schemes,
             ideTemplateMacros: nil,
-            additionalFiles: []
+            additionalFiles: [],
+            resourceSynthesizers: []
         )
     }