+++ Tests/TuistIntegrationTests/Generator/TestModelGenerator.swift
@@ -117,7 +117,8 @@ final class TestModelGenerator {
             packages: packages,
             schemes: schemes,
             ideTemplateMacros: nil,
-            additionalFiles: createAdditionalFiles(path: path)
+            additionalFiles: createAdditionalFiles(path: path),
+            resourceSynthesizers: []
         )
     }