+++ Tests/TuistPluginTests/PluginServiceTests.swift
@@ -41,7 +41,8 @@ final class PluginServiceTests: TuistTestCase {
     }
 
     func test_loadPlugins_WHEN_localHelpers() throws {
-        let pluginPath = AbsolutePath("/path/to/Plugin")
+        // Given
+        let pluginPath = try temporaryPath().appending(component: "Plugin")
         let pluginName = "TestPlugin"
 
         manifestLoader.loadConfigStub = { _ in
@@ -52,18 +53,23 @@ final class PluginServiceTests: TuistTestCase {
             ProjectDescription.Plugin(name: pluginName)
         }
 
-        fileHandler.stubExists = { _ in
-            true
-        }
-
         let config = mockConfig(plugins: [TuistGraph.PluginLocation.local(path: pluginPath.pathString)])
+
+        try fileHandler.createFolder(
+            pluginPath.appending(component: Constants.helpersDirectoryName)
+        )
+
+        // When
         let plugins = try subject.loadPlugins(using: config)
+
+        // Then
         let expectedHelpersPath = pluginPath.appending(component: Constants.helpersDirectoryName)
         let expectedPlugins = Plugins.test(projectDescriptionHelpers: [.init(name: pluginName, path: expectedHelpersPath, location: .local)])
         XCTAssertEqual(plugins, expectedPlugins)
     }
 
     func test_loadPlugins_WHEN_gitHelpers() throws {
+        // Given
         let pluginGitUrl = "https://url/to/repo.git"
         let pluginGitId = "1.0.0"
         let pluginFingerprint = "\(pluginGitUrl)-\(pluginGitId)".md5
@@ -78,14 +84,75 @@ final class PluginServiceTests: TuistTestCase {
             ProjectDescription.Plugin(name: pluginName)
         }
 
-        fileHandler.stubExists = { _ in
-            true
+        try fileHandler.createFolder(cachedPluginPath.appending(component: Constants.helpersDirectoryName))
+
+        let config = mockConfig(plugins: [TuistGraph.PluginLocation.gitWithTag(url: pluginGitUrl, tag: pluginGitId)])
+
+        // When
+        let plugins = try subject.loadPlugins(using: config)
+
+        // Then
+        let expectedHelpersPath = cachedPluginPath.appending(component: Constants.helpersDirectoryName)
+        let expectedPlugins = Plugins.test(projectDescriptionHelpers: [.init(name: pluginName, path: expectedHelpersPath, location: .remote)])
+        XCTAssertEqual(plugins, expectedPlugins)
+    }
+
+    func test_loadPlugins_when_localResourceSynthesizer() throws {
+        // Given
+        let pluginPath = try temporaryPath()
+        let pluginName = "TestPlugin"
+        let resourceTemplatesPath = pluginPath.appending(components: "ResourceSynthesizers")
+
+        try makeDirectories(resourceTemplatesPath)
+
+        manifestLoader.loadConfigStub = { _ in
+            .test(plugins: [.local(path: .relativeToRoot(pluginPath.pathString))])
+        }
+
+        manifestLoader.loadPluginStub = { _ in
+            ProjectDescription.Plugin(name: pluginName)
+        }
+
+        let config = mockConfig(plugins: [TuistGraph.PluginLocation.local(path: pluginPath.pathString)])
+
+        // When
+        let plugins = try subject.loadPlugins(using: config)
+        let expectedPlugins = Plugins.test(
+            resourceSynthesizers: [
+                PluginResourceSynthesizer(name: pluginName, path: resourceTemplatesPath),
+            ]
+        )
+        XCTAssertEqual(plugins, expectedPlugins)
+    }
+
+    func test_loadPlugins_when_remoteResourceSynthesizer() throws {
+        // Given
+        let pluginGitUrl = "https://url/to/repo.git"
+        let pluginGitId = "1.0.0"
+        let pluginFingerprint = "\(pluginGitUrl)-\(pluginGitId)".md5
+        let cachedPluginPath = environment.cacheDirectory.appending(components: Constants.pluginsDirectoryName, pluginFingerprint)
+        let pluginName = "TestPlugin"
+        let resourceTemplatesPath = cachedPluginPath.appending(components: "ResourceSynthesizers")
+
+        try makeDirectories(resourceTemplatesPath)
+
+        manifestLoader.loadConfigStub = { _ in
+            .test(plugins: [ProjectDescription.PluginLocation.git(url: pluginGitUrl, tag: pluginGitId)])
+        }
+
+        manifestLoader.loadPluginStub = { _ in
+            ProjectDescription.Plugin(name: pluginName)
         }
 
         let config = mockConfig(plugins: [TuistGraph.PluginLocation.gitWithTag(url: pluginGitUrl, tag: pluginGitId)])
+
+        // When
         let plugins = try subject.loadPlugins(using: config)
-        let expectedHelpersPath = cachedPluginPath.appending(component: Constants.helpersDirectoryName)
-        let expectedPlugins = Plugins.test(projectDescriptionHelpers: [.init(name: pluginName, path: expectedHelpersPath, location: .remote)])
+        let expectedPlugins = Plugins.test(
+            resourceSynthesizers: [
+                PluginResourceSynthesizer(name: pluginName, path: resourceTemplatesPath),
+            ]
+        )
         XCTAssertEqual(plugins, expectedPlugins)
     }