+++ projects/tuist/fixtures/app_with_plugins/Source/AppDelegate.swift
@@ -6,6 +6,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
         window = UIWindow(frame: UIScreen.main.bounds)
+        
+        print(Strings.hello)
+        #if canImport(Lottie)
+        print(AnimationAsset.allAnimations.everythingBagel)
+        #endif
+        
         let viewController = UIViewController()
         viewController.view.backgroundColor = .white
         viewController.view.addSubview(TestView())